% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_opl.R
\name{pin_opl}
\alias{pin_opl}
\title{Utility function to pin the csv of the latest OpenPowerLifting data
with the Pins 1.0.0 API.}
\usage{
pin_opl(board, ipf = TRUE, name = NULL)
}
\arguments{
\item{board}{the Board object from {pins} to attach the data to}

\item{ipf}{LOGICAL, IPF-affiliated competitons only, default TRUE.}

\item{name}{optional STRING for the name for the pin. Defaults to "opl-ipf" or "opl-all"}
}
\value{
Same as pin_write
}
\description{
Utility function to pin the csv of the latest OpenPowerLifting data
with the Pins 1.0.0 API.
}
