% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_opl.R
\name{sql_opl}
\alias{sql_opl}
\title{Utility function to pin the csv of the latest OpenPowerLifting data
with the Pins 1.0.0 API.}
\usage{
sql_opl(ipf = TRUE, name = NULL, location = "~/.local/share/opl")
}
\arguments{
\item{ipf}{LOGICAL, IPF-affiliated competitons only, default TRUE.}

\item{name}{optional STRING for the name for the table. Defaults to "opl-ipf" or "opl-all"}

\item{location}{optional string for the location to store the SQLite DB.}
}
\value{
name of the table, invisibly
}
\description{
Utility function to pin the csv of the latest OpenPowerLifting data
with the Pins 1.0.0 API.
}
