% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{pin_opl_full}
\alias{pin_opl_full}
\title{Utility function for pinning the latest full OPL extract.}
\usage{
pin_opl_full(name = NULL)
}
\arguments{
\item{name}{optional STRING for the name for the pin. Defaults to "opl-ipf" or "opl-all"}

\item{board}{the Board object from {pins} to attach the data to. Defaults to pins::board_local().}
}
\value{
Same as pin_write
}
\description{
Utility function for pinning the latest full OPL extract.
}
